refactor: AuthService에서만 사용되는 LoginTokens 내부 레코드 클래스로 변경
